requirements.txt
Flask==3.0.0
Pillow==10.2.0
qrcode==7.4.2
